# Improved UnorderedOffsetManager