Signed-off-by: Francesco Guardiani <francescoguard@gmail.com>

## Proposed Changes

- Modified a bit the benchmark from #640 to make sure we don't measure the record allocation time (which is quite expensive in longer benchmarks and dominated over the code under test), but now we record the `UnorderdOffsetManager` allocation (as noted in the JMH javadocs, it's better to avoid using `Level.Invocation` for the state when the bench length is short
- Rewrote the unordered offset manager to be more efficient

## Numbers

Master + https://github.com/knative-sandbox/eventing-kafka-broker/pull/648/commits/1df960a732bc5c76d74ada2fe50ee708606e193d:

```
Benchmark                                                Mode  Cnt       Score      Error  Units
UnorderedOffsetManagerBenchmark.benchmarkMixedABit      thrpt   10   23918,942 ± 1208,231  ops/s
UnorderedOffsetManagerBenchmark.benchmarkOrdered        thrpt   10       7,437 ±    0,130  ops/s
UnorderedOffsetManagerBenchmark.benchmarkRealisticCase  thrpt   10  150935,199 ± 3552,501  ops/s
UnorderedOffsetManagerBenchmark.benchmarkReverseOrder   thrpt   10       2,745 ±    0,103  ops/s
```

This PR:

```
Benchmark                                                Mode  Cnt       Score       Error  Units
UnorderedOffsetManagerBenchmark.benchmarkMixedABit      thrpt   10   61381,474 ±  1125,869  ops/s
UnorderedOffsetManagerBenchmark.benchmarkOrdered        thrpt   10       5,453 ±     0,026  ops/s
UnorderedOffsetManagerBenchmark.benchmarkRealisticCase  thrpt   10  481574,607 ± 16121,914  ops/s
UnorderedOffsetManagerBenchmark.benchmarkReverseOrder   thrpt   10       7,807 ±     0,145  ops/s
```

Note: the realistic benchmarks are `benchmarkMixedABit` and `benchmarkRealisticCase`, the other 2 are extreme cases where probably there are other factors dominating over the code under test